requirements.txt
litellm>=1.57.0
PyGithub>=2.1.1
google-api-python-client>=2.110.0
youtube-transcript-api>=0.6.2
feedparser>=6.0.11
requests>=2.31.0
python-dotenv>=1.0.0
PyYAML>=6.0.3
jinja2>=3.1.0
markdown>=3.5
pytz>=2025.2
tenacity>=8.5.0
